refactor(Docker): modularize Docker command execution and enhance test coverage
Introduce `BaseCommand` and `RunCommand` traits to encapsulate common Docker command functionalities, improving code modularity and reusability. This refactoring extracts command execution logic from `DockerClient` into separate traits, reducing duplication and enhancing maintainability. The `ContainerId` class is added to represent Docker container IDs, ensuring type safety and validation. Tests are updated to reflect these changes, with new tests added for `RunCommand` to ensure comprehensive coverage. This modular approach facilitates easier future extensions and modifications to Docker command handling.
