+++ src/Docker/DockerClient.php
@@ -3,14 +3,14 @@
 namespace Testcontainers\Docker;
 
 use Symfony\Component\Process\Process;
+use Testcontainers\Docker\Command\BaseCommand;
+use Testcontainers\Docker\Command\RunCommand;
 use Testcontainers\Docker\Exception\DockerException;
 use Testcontainers\Docker\Exception\InvalidValueException;
 use Testcontainers\Docker\Exception\NoSuchContainerException;
 use Testcontainers\Docker\Exception\NoSuchObjectException;
-use Testcontainers\Docker\Exception\PortAlreadyAllocatedException;
 
 use function Testcontainers\array_flatten;
-use function Testcontainers\kebab;
 
 /**
  * A client for interacting with Docker.
@@ -20,239 +20,8 @@ use function Testcontainers\kebab;
  */
 class DockerClient
 {
-    /**
-     * The command used to interact with Docker.
-     *
-     * @var string The Docker command, default is 'docker'.
-     */
-    private $command = 'docker';
-
-    /**
-    * Global options for Docker commands.
-    *
-    * These options are applied to all Docker commands executed by this client.
-    *
-    * @var array
-    */
-    private $options = [];
-
-    /**
-     * The working directory for Docker commands.
-     *
-     * This directory is used as the current working directory for all Docker commands
-     * executed by this client. If set to null, the current PHP process directory is used.
-     *
-     * @var null|string The working directory path or null if not set.
-     */
-    private $cwd = null;
-
-    /**
-     * Environment variables for Docker commands.
-     *
-     * These variables are passed to the Docker process when executing commands.
-     * If set to null, the current environment variables of the PHP process are used.
-     *
-     * @var array An associative array of environment variables or null if not set.
-     */
-    private $env = [];
-
-    /**
-     * The input for Docker commands.
-     *
-     * This can be a stream resource, scalar value, `\Traversable`, or null if no input is provided.
-     *
-     * @var mixed|null
-     */
-    private $input = null;
-
-    /**
-     * The timeout for Docker commands.
-     *
-     * This value represents the maximum time in seconds that a Docker command is allowed to run.
-     * If set to null, the command will run indefinitely without timing out.
-     *
-     * @var int|float|null The timeout in seconds, or null to disable the timeout.
-     */
-    private $timeout = 60;
-
-    /**
-     * Options for the `proc_open` function.
-     *
-     * These options are used when creating a new process using the `proc_open` function.
-     * Refer to the PHP documentation for `proc_open` for more details on available options.
-     *
-     * @var array
-     */
-    private $proc_options = [];
-
-    /**
-     * Set the Docker command path.
-     *
-     * This method allows you to specify the path to the Docker command
-     * that will be used by this client. By default, it is set to 'docker'.
-     *
-     * @param string $command The path to the Docker command.
-     * @return self
-     */
-    public function withCommand($command)
-    {
-        $this->command = $command;
-
-        return $this;
-    }
-
-    /**
-     * Set global Docker options.
-     *
-     * This method allows you to specify global options that will be applied to all Docker commands
-     * executed by this client. These options can include flags and parameters that modify the behavior
-     * of Docker commands.
-     *
-     * @param array $options An associative array of Docker global options.
-     * @return self
-     */
-    public function withGlobalOptions($options)
-    {
-        $this->options = $options;
-
-        return $this;
-    }
-
-    /**
-     * Set the working directory for Docker commands.
-     *
-     * This method sets the directory that will be used as the current working directory
-     * for all Docker commands executed by this client. If not set, the current PHP process
-     * directory will be used.
-     *
-     * @param null|string $cwd The path to the working directory.
-     * @return self
-     */
-    public function withCwd($cwd)
-    {
-        $this->cwd = $cwd;
-
-        return $this;
-    }
-
-    /**
-     * Set environment variables for Docker commands.
-     *
-     * This method allows you to specify environment variables that will be passed
-     * to the Docker process when executing commands. If not set, the current
-     * environment variables of the PHP process will be used.
-     *
-     * @param array $env An associative array of environment variables.
-     * @return self
-     */
-    public function withEnv($env)
-    {
-        $this->env = $env;
-
-        return $this;
-    }
-
-    /**
-     * Set the input for Docker commands.
-     *
-     * This method allows you to specify the input that will be passed to the Docker process
-     * when executing commands. The input can be a stream resource, a scalar value, a `\Traversable`,
-     * or `null` if no input is provided.
-     *
-     * @param null|mixed $input The input for the Docker process.
-     * @return self
-     */
-    public function withInput($input)
-    {
-        $this->input = $input;
-
-        return $this;
-    }
-
-    /**
-     * Set the timeout for Docker commands.
-     *
-     * This method allows you to specify the maximum time in seconds that a Docker command is allowed to run.
-     * If set to `null`, the command will run indefinitely without timing out.
-     *
-     * @param int|float|null $timeout The timeout in seconds, or `null` to disable the timeout.
-     * @return self
-     */
-    public function withTimeout($timeout)
-    {
-        $this->timeout = $timeout;
-
-        return $this;
-    }
-
-    /**
-     * Set options for the `proc_open` function.
-     *
-     * This method allows you to specify options that will be used when creating a new process
-     * using the `proc_open` function. Refer to the PHP documentation for `proc_open` for more details
-     * on available options.
-     *
-     * @param array $proc_options An associative array of options for `proc_open`.
-     * @return self
-     */
-    public function withProcOptions($proc_options)
-    {
-        $this->proc_options = $proc_options;
-
-        return $this;
-    }
-
-    /**
-     * Create and run a new container from a Docker image.
-     *
-     * This method wraps the `docker run` command to create and run a new container from a specified Docker image.
-     *
-     * @param string $image The name of the Docker image to use.
-     * @param string|null $command The command to run inside the container (optional).
-     * @param array|null $args The arguments for the command (optional).
-     * @param array $options Additional options for the Docker command.
-     * @return DockerRunOutput|DockerRunWithDetachOutput The output of the Docker run command. If the `detach` option is set to `true`, a `DockerRunWithDetachOutput` object is returned.
-     *
-     * @throws PortAlreadyAllocatedException If the specified port is already allocated.
-     * @throws DockerException If the Docker command fails.
-     */
-    public function run($image, $command = null, $args = null, $options = [])
-    {
-        $commandline = array_filter(array_flatten([
-            $this->command,
-            $this->arrayToArgs($this->options),
-            'run',
-            $this->arrayToArgs($options),
-            $image,
-            $command,
-            $args,
-        ]));
-        $process = new Process(
-            $commandline,
-            $this->cwd,
-            $this->env,
-            $this->input,
-            $this->timeout,
-            $this->proc_options
-        );
-        $process->run();
-
-        if (!$process->isSuccessful()) {
-            $stderr = $process->getErrorOutput();
-            if (PortAlreadyAllocatedException::match($stderr)) {
-                throw new PortAlreadyAllocatedException($process);
-            }
-            throw new DockerException($process);
-        }
-
-        if (isset($options['detach']) && $options['detach'] === true) {
-            $stdout = $process->getOutput();
-            $containerId = trim($stdout);
-            return new DockerRunWithDetachOutput($process, $containerId);
-        } else {
-            return new DockerRunOutput($process);
-        }
-    }
+    use BaseCommand;
+    use RunCommand;
 
     /**
      * Inspect a Docker container.
@@ -503,47 +272,6 @@ class DockerClient
         }
 
         return new DockerNetworkCreateOutput($process);
-    }
-
-    /**
-     * Convert array to command line arguments
-     *
-     * @param array $options command line options (key-value pairs)
-     * @return array
-     */
-    private function arrayToArgs($options)
-    {
-        $result = [];
-        foreach ($options as $key => $value) {
-            $key = kebab($key);
-            if ($value === null) {
-                continue;
-            }
-            if ($value === false) {
-                continue;
-            }
-            if ($value === true) {
-                $result[] = "--$key";
-                continue;
-            }
-            if (is_scalar($value)) {
-                $result[] = "--$key";
-                $result[] = $value;
-                continue;
-            }
-            if (is_array($value)) {
-                foreach ($value as $k => $v) {
-                    $result[] = "--$key";
-                    if (is_string($k)) {
-                        $result[] = "$k=$v";
-                    } elseif (is_scalar($v) && !is_bool($v)) {
-                        $result[] = $v;
-                    }
-                }
-            }
-        }
-
-        return $result;
     }
 
     public function __clone()
