+++ tests/Unit/Docker/DockerClientTest.php
@@ -9,109 +9,13 @@ use Testcontainers\Docker\DockerInspectOutput;
 use Testcontainers\Docker\DockerLogsOutput;
 use Testcontainers\Docker\DockerNetworkCreateOutput;
 use Testcontainers\Docker\DockerProcessStatusOutput;
-use Testcontainers\Docker\DockerRunOutput;
-use Testcontainers\Docker\DockerRunWithDetachOutput;
 use Testcontainers\Docker\DockerStopOutput;
 use Testcontainers\Docker\Exception\NoSuchContainerException;
-use Testcontainers\Docker\Exception\PortAlreadyAllocatedException;
 use Testcontainers\Testcontainers;
 use Tests\Images\DinD;
 
 class DockerClientTest extends TestCase
 {
-    public function testRun()
-    {
-        $client = new DockerClient();
-        $output = $client->run('alpine:latest', 'echo', ['Hello, World!']);
-
-        $this->assertInstanceOf(DockerRunOutput::class, $output);
-        $this->assertSame(0, $output->getExitCode());
-        $this->assertSame("Hello, World!\n", $output->getOutput());
-    }
-
-    public function testRunWithDetach()
-    {
-        $client = new DockerClient();
-        $output = $client->run('alpine:latest', 'echo', ['Hello, World!'], [
-            'detach' => true,
-        ]);
-
-        $this->assertInstanceOf(DockerRunWithDetachOutput::class, $output);
-        $this->assertSame(0, $output->getExitCode());
-        $this->assertNotEmpty($output->getContainerId());
-    }
-
-    public function testRunWithTrueOptions()
-    {
-        $client = new DockerClient();
-        $output = $client->run('alpine:latest', 'echo', ['Hello, World!'], [
-            'quiet' => true,
-        ]);
-
-        $this->assertInstanceOf(DockerRunOutput::class, $output);
-        $this->assertSame(0, $output->getExitCode());
-        $this->assertSame("Hello, World!\n", $output->getOutput());
-    }
-
-    public function testRunWithFalseOptions()
-    {
-        $client = new DockerClient();
-        $output = $client->run('alpine:latest', 'echo', ['Hello, World!'], [
-            'detach' => false,
-        ]);
-
-        $this->assertInstanceOf(DockerRunOutput::class, $output);
-        $this->assertSame(0, $output->getExitCode());
-        $this->assertSame("Hello, World!\n", $output->getOutput());
-    }
-
-    public function testRunWithArrayOptions()
-    {
-        $client = new DockerClient();
-        $output = $client->run('alpine:latest', 'echo', ['Hello, World!'], [
-            'publish' => ['38621:80', '38622:443'],
-        ]);
-
-        $this->assertInstanceOf(DockerRunOutput::class, $output);
-        $this->assertSame(0, $output->getExitCode());
-        $this->assertSame("Hello, World!\n", $output->getOutput());
-    }
-
-    public function testRunWithObjectOptions()
-    {
-        $client = new DockerClient();
-        $output = $client->run('alpine:latest', 'printenv', ['BAR'], [
-            'env' => [
-                'FOO' => 'foo',
-                'BAR' => 'bar',
-            ],
-        ]);
-
-        $this->assertInstanceOf(DockerRunOutput::class, $output);
-        $this->assertSame(0, $output->getExitCode());
-        $this->assertSame("bar\n", $output->getOutput());
-    }
-
-    public function testRunWithPortConflict()
-    {
-        $this->expectException(PortAlreadyAllocatedException::class);
-
-        $instance = Testcontainers::run(DinD::class);
-
-        $client = new DockerClient();
-        $client->withGlobalOptions([
-            'host' => 'tcp://' . $instance->getHost() . ':' . $instance->getMappedPort(2375),
-        ]);
-        $client->run('nginx:1.27.2', null, null, [
-            'detach' => true,
-            'publish' => ['38793:80'],
-        ]);
-        $client->run('nginx:1.27.2', null, null, [
-            'detach' => true,
-            'publish' => ['38793:80'],
-        ]);
-    }
-
     public function testInspect()
     {
         $client = new DockerClient();
@@ -181,7 +85,7 @@ class DockerClientTest extends TestCase
         $client->withGlobalOptions([
             'host' => 'tcp://' . $instance->getHost() . ':' . $instance->getMappedPort(2375),
         ]);
-        $output = $client->run('jpetazzo/clock:latest', null, null, [
+        $output = $client->run('jpetazzo/clock:latest', null, [], [
             'detach' => true,
         ]);
 
@@ -201,7 +105,7 @@ class DockerClientTest extends TestCase
         $client->withGlobalOptions([
             'host' => 'tcp://' . $instance->getHost() . ':' . $instance->getMappedPort(2375),
         ]);
-        $output = $client->run('jpetazzo/clock:latest', null, null, [
+        $output = $client->run('jpetazzo/clock:latest', null, [], [
             'detach' => true,
         ]);
